Extracted test code from arduino.py
